jf_out: row/column affects held keys correctly, patcher: connect, fix random load fail in tune
