+++ lib/ui/norns.lua
@@ -99,20 +99,6 @@ local function App()
 
     return function()
         if eggs.view_focus == eggs.NORMAL then 
-            -- _text{ x = x[1], y = y[1], text = 'this is eggs' }
-            
-            -- do
-            --     local id = 'target_'..eggs.track_focus
-            --     _dest.enc{
-            --         n = 2, state = crops.of_param(id),
-            --         max = #params:lookup_param(id).options,
-            --     }
-            --     _dest.screen{
-            --         x = x[1], y = y[2], flow = 'down',
-            --         text = { ['destination'] = params:string(id) },
-            --     }
-            -- end
-
             _pages[eggs.track_focus]()
         else
             _tuning{ track = eggs.track_focus, view = eggs.view_focus }
